Moved InvalidKey error into the backpressure module.
