+++ runtime/swimos_runtime/src/agent/task/mod.rs
@@ -22,7 +22,7 @@ use crate::agent::store::StoreInitError;
 use crate::agent::task::links::TriggerUnlink;
 use crate::agent::task::sender::LaneSendError;
 use crate::agent::task::write_fut::SpecialAction;
-use crate::error::InvalidKey;
+use crate::backpressure::InvalidKey;
 use crate::timeout_coord::{self, VoteResult};
 
 use self::external_links::{LinksTaskState, NoReport};
