+++ runtime/swimos_runtime/src/agent/task/remotes/uplink/mod.rs
@@ -29,9 +29,8 @@ use crate::{
     },
     backpressure::{
         recon::MapOperationReconEncoder, BackpressureStrategy, MapBackpressure, SupplyBackpressure,
-        ValueBackpressure,
+        ValueBackpressure, InvalidKey
     },
-    error::InvalidKey,
 };
 
 #[cfg(test)]
