+++ runtime/swimos_runtime/src/backpressure/map_queue/mod.rs
@@ -20,9 +20,7 @@ use std::{
 use bytes::{BufMut, BytesMut};
 use swimos_agent_protocol::MapOperation;
 
-use crate::error::InvalidKey;
-
-use super::{key::ReconKey, RawMapOperation, RawMapOperationMut};
+use super::{key::ReconKey, RawMapOperation, RawMapOperationMut, InvalidKey};
 
 #[cfg(test)]
 mod tests;
